refactor: move common block components to blocks
